requirements/common.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-Bootstrap==3.3.7.0
Flask-Script==2.0.5
Flask-Migrate==2.0.0
six==1.10.0